+++ Project.toml
@@ -9,6 +9,7 @@ Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
 DataAPI = "9a962f9c-6df0-11e9-0e5d-c546b8b5ee8a"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 FLoops = "cc61a311-1640-44b5-9fba-1b764f453329"
+MLCore = "c2834f40-e789-41da-a90e-33b280584a8c"
 NNlib = "872c559c-99b0-510c-b3b7-b6c96a88d5cd"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 ShowCases = "605ecd9f-84a6-4c9e-81e2-4798472b76a3"
@@ -24,6 +25,7 @@ Compat = "4.2"
 DataAPI = "1.0"
 DelimitedFiles = "1.0"
 FLoops = "0.2"
+MLCore = "1.0.0"
 NNlib = "0.8, 0.9"
 ShowCases = "0.1"
 SimpleTraits = "0.9"
