+++ src/slidingwindow.jl
@@ -38,7 +38,7 @@ To actually get a copy of the data at some window use indexing or [`getobs`](@re
 When indexing the data is accessed as `getobs(data, idxs)`, with `idxs` an appropriate range of indexes.
 ```jldoctest
 julia> s = slidingwindow(11:30, size=6)
-slidingwindow(10:30, size=6, stride=1)
+slidingwindow(11:30, size=6, stride=1)
 
 julia> s[1]  # == getobs(data, 1:6)
 11:16
@@ -53,7 +53,7 @@ By default the stride is equal to 1.
 
 ```jldoctest
 julia> s = slidingwindow(11:30, size=6, stride=3)
-slidingwindow(1:20, size=6, stride=3)
+slidingwindow(11:30, size=6, stride=3)
 
 julia> for w in s; println(w); end
 11:16
